requirements.txt
flask
pymongo
python-dotenv
flask-wtf
email-validator
dnspython
requests
cloudinary
Pillow